requirements.txt
itk-elastix
matplotlib
numpy
pandas
pydicom
scikit-image
tensorflow
xarray
tqdm
pre-commit
scipy
itk
PyYAML
einops
hydra-core
timm
torch
pytest
black
dipy
opencv-python
scikit-learn
mayavi
PyWavelets
pre-commit